+++ server/src/main/java/org/allaymc/server/block/component/BlockHopperBaseComponentImpl.java
@@ -1,19 +1,26 @@
 package org.allaymc.server.block.component;
 
 import org.allaymc.api.block.BlockBehavior;
+import org.allaymc.api.block.component.BlockBlockEntityHolderComponent;
 import org.allaymc.api.block.data.BlockFace;
 import org.allaymc.api.block.dto.Block;
 import org.allaymc.api.block.dto.PlayerInteractInfo;
 import org.allaymc.api.block.property.type.BlockPropertyTypes;
 import org.allaymc.api.block.type.BlockState;
 import org.allaymc.api.block.type.BlockType;
+import org.allaymc.api.blockentity.interfaces.BlockEntityHopper;
 import org.allaymc.api.world.Dimension;
+import org.allaymc.server.component.annotation.Dependency;
 import org.joml.Vector3ic;
 
 /**
  * @author ClexaGod | daoge_cmd
  */
 public class BlockHopperBaseComponentImpl extends BlockBaseComponentImpl {
+
+    @Dependency
+    private BlockBlockEntityHolderComponent<BlockEntityHopper> blockEntityHolderComponent;
+
     public BlockHopperBaseComponentImpl(BlockType<? extends BlockBehavior> blockType) {
         super(blockType);
     }
@@ -44,5 +51,19 @@ public class BlockHopperBaseComponentImpl extends BlockBaseComponentImpl {
         if (powered != currentlyDisabled) {
             block.updateBlockProperty(BlockPropertyTypes.TOGGLE_BIT, powered);
         }
+    }
+
+    @Override
+    public boolean hasComparatorInputOverride() {
+        return true;
+    }
+
+    @Override
+    public int getComparatorInputOverride(Block block) {
+        var hopper = blockEntityHolderComponent.getBlockEntity(block.getPosition());
+        if (hopper == null) {
+            return 0;
+        }
+        return hopper.getContainer().calculateComparatorSignal();
     }
 }